services/ai_interviewer/requirements.txt
Flask
Flask-CORS
python-dotenv
pymongo
groq
requests